Change decorator to use regex
The decorator becomes very inefficient if it looks at the text in the previous contentBlock.
A better strategy is to decorate strings preceding or following colons. Although this means that
the coloring will break down if the text is not written properly, such as entering line breaks
just before the colon or entering line breaks after the colon when the value is not an array or
object but just a string. Maybe I will handle these as special cases that look at previous/next
contentblock later on. But for now this is efficient.